+++ client/src/components/DependencyDetailsForm/DependencyDetailsForm.js
@@ -1,32 +1,28 @@
 import {useState} from 'react';
-import {useHistory} from 'react-router-dom';
-import {Form, Button, Spin, Typography, Select} from 'antd';
+import {Form, Button, Spin, Select, Row, Col, Typography} from 'antd';
+import {useFormik} from 'formik';
 import axios from 'axios';
-import {useProjectContext} from '../../providers/Project';
 import css from './dependencyDetailsForm.module.css';
 
 const {Title} = Typography;
 const {Option} = Select;
 
-const DependencyDetailsForm = ({serviceInfo, setServiceInfo, dependencyOptions}) => {
+const DependencyDetailsForm = ({serviceInfo, setServiceInfo, dependencyOptions, onSuccess}) => {
   const serviceId = serviceInfo._id;
-  let history = useHistory();
   const [submitting, setSubmitting] = useState(false);
   const [error, setError] = useState(null);
+  const formik = useFormik({
+    initialValues: serviceInfo,
+    onSubmit: (values) => onSave(values),
+  });
 
-  const onFinishFailed = (errorInfo) => {
-    console.log('Failed:', errorInfo);
-  };
-
-  const handleSubmit = async (values) => {
+  const onSave = async (values) => {
     try {
       setSubmitting(true);
-      const {data} = await axios.put(`http://localhost:8080/service/${serviceId}`, {
-        ...values,
-      });
-      setServiceInfo(data);
+      const {data} = await axios.put(`http://localhost:8080/service/${serviceId}`, values);
+      setServiceInfo(values);
       setSubmitting(false);
-      history.push(`/service/${serviceId}`);
+      onSuccess();
     } catch (e) {
       console.log(e);
       setError(e);
@@ -37,38 +33,23 @@ const DependencyDetailsForm = ({serviceInfo, setServiceInfo, dependencyOptions})
     return null;
   }
 
-  console.log(dependencyOptions);
-
   return (
     <div>
       <div className={css.spinnerContainer}>
         <Spin spinning={submitting} tip="Loading..." />
       </div>
       {!submitting && (
-        <>
-          <Form
-            layout="vertical"
-            name="dependencyDetailsForm"
-            initialValues={{
-              dependency: serviceInfo.dependency,
-            }}
-            onFinish={handleSubmit}
-            onFinishFailed={onFinishFailed}>
-            <Form.Item
-              label="Dependencies"
-              name="dependency"
-              required={true}
-              rules={[
-                {
-                  required: true,
-                  message: 'Cannot be blank',
-                },
-              ]}>
+        <Form>
+          <Row gutter={[20, 20]}>
+            <Col xs={24}>
+              <Title level={5}>Dependencies</Title>
               <Select
+                style={{width: '100%'}}
+                name="dependency"
                 mode="multiple"
-                // allowClear
+                value={formik.values.dependency}
                 placeholder="Please select dependencies"
-                onChange={(e) => console.log(e)}>
+                onChange={(newValues) => formik.setFieldValue('dependency', newValues)}>
                 {dependencyOptions?.map(({_id, name}) => {
                   if (serviceId !== _id) {
                     return (
@@ -79,15 +60,14 @@ const DependencyDetailsForm = ({serviceInfo, setServiceInfo, dependencyOptions})
                   }
                 })}
               </Select>
-            </Form.Item>
-
-            <Form.Item>
-              <Button type="primary" htmlType="submit">
+            </Col>
+            <Col xs={24}>
+              <Button type="primary" onClick={() => formik.handleSubmit(formik.values)}>
                 Submit
               </Button>
-            </Form.Item>
-          </Form>
-        </>
+            </Col>
+          </Row>
+        </Form>
       )}
     </div>
   );